Fix type incompatibility
the type of the frame_node name field was originally defined as
swaddr_t without serious consideration. The `name' field is mostly
used in the nemu monitor and it should be combined to the platform
that the nemu is built. But nemu itself simulate a IA32 platform
and swaddr_t is designed for this simulated one. Therefore if we
build nemu under x86_64, we will cast 64bit addr to 32bit integer!

BTW, the uint64_t cannot be considered consistent to %lld format by
the gcc under x86_64, so I use the coersion cast (long long unsigned
int).