Back up the cursor in stack to avoid hazard
The puts() in timer callback and the normal print function in
executing loop both modify the cursor (line, col) but does not
recognize each other, resulting the cursor excceds text memory
space.

We use local variables to store the cursor when entering the basic
printf function and update it just before return.